[AutoBindParameter] Decouple new package
