+++ composer.json
@@ -47,7 +47,7 @@
         "doctrine/persistence": "^1.3.3",
         "phpstan/phpstan": "^0.12",
         "phpunit/phpunit": "^8.4",
-        "rector/rector": "^0.6.2",
+        "rector/rector": "^0.6",
         "symfony/framework-bundle": "^4.4|^5.0",
         "symfony/templating": "^4.4|^5.0",
         "symfony/translation": "^4.4|^5.0",
@@ -56,6 +56,7 @@
     },
     "autoload": {
         "psr-4": {
+            "Symplify\\AutoBindParameter\\": "packages/AutoBindParameter/src",
             "Symplify\\Autodiscovery\\": "packages/Autodiscovery/src",
             "Symplify\\AutowireArrayParameter\\": "packages/AutowireArrayParameter/src",
             "Symplify\\ChangelogLinker\\": "packages/ChangelogLinker/src",
@@ -95,6 +96,7 @@
             "packages/Autodiscovery/tests/Source/KernelProjectDir"
         ],
         "psr-4": {
+            "Symplify\\AutoBindParameter\\Tests\\": "packages/AutoBindParameter/tests",
             "Symplify\\Autodiscovery\\Tests\\": "packages/Autodiscovery/tests",
             "Symplify\\AutowireArrayParameter\\Tests\\": "packages/AutowireArrayParameter/tests",
             "Symplify\\ChangelogLinker\\Tests\\": "packages/ChangelogLinker/tests",
@@ -140,7 +142,8 @@
         "symplify/smart-file-system": "7.1",
         "symplify/set-config-resolver": "7.1",
         "symplify/statie": "7.1",
-        "symplify/autowire-array-parameter": "7.1"
+        "symplify/autowire-array-parameter": "7.1",
+        "symplify/auto-bind-parameter": "7.1"
     },
     "scripts": {
         "complete-check": [
