+++ packages/AutoBindParameter/.gitattributes
@@ -1,0 +1,5 @@
+.gitattributes export-ignore
+.gitignore export-ignore
+.travis.yml export-ignore
+phpunit.xml export-ignore
+*.md export-ignore
