+++ packages/AutoBindParameter/.travis.yml
@@ -1,0 +1,17 @@
+language: php
+
+matrix:
+    include:
+        - php: 7.2
+          env: COMPOSER_FLAGS="--prefer-lowest"
+        - php: 7.3
+        - php: 7.4
+
+install:
+    - composer update --prefer-source $COMPOSER_FLAGS
+
+script:
+    - vendor/bin/phpunit
+
+notifications:
+    email: false
